// src/api/models/ComponenteDisponibleResponseListApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ComponenteDisponibleResponse } from './ComponenteDisponibleResponse';
export type ComponenteDisponibleResponseListApiResponse = {
    success?: boolean;
    data?: Array<ComponenteDisponibleResponse> | null;
    message?: string | null;
    timestamp?: string;
    errors?: Array<string> | null;
};