// src/api/models/EstadoComponenteIngresoResponseApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { EstadoComponenteIngresoResponse } from './EstadoComponenteIngresoResponse';
export type EstadoComponenteIngresoResponseApiResponse = {
    success?: boolean;
    data?: EstadoComponenteIngresoResponse;
    message?: string | null;
    timestamp?: string;
    errors?: Array<string> | null;
};