// src/api/models/InformacionPreviaOrdenResponseApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { InformacionPreviaOrdenResponse } from './InformacionPreviaOrdenResponse';
export type InformacionPreviaOrdenResponseApiResponse = {
    success?: boolean;
    data?: InformacionPreviaOrdenResponse;
    message?: string | null;
    timestamp?: string;
    errors?: Array<string> | null;
};