// src/api/models/InspectorOrdenResponseApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { InspectorOrdenResponse } from './InspectorOrdenResponse';
export type InspectorOrdenResponseApiResponse = {
    success?: boolean;
    data?: InspectorOrdenResponse;
    message?: string | null;
    timestamp?: string;
    errors?: Array<string> | null;
};