// src/api/models/OrdenDetalladaResponseApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { OrdenDetalladaResponse } from './OrdenDetalladaResponse';
export type OrdenDetalladaResponseApiResponse = {
    success?: boolean;
    data?: OrdenDetalladaResponse;
    message?: string | null;
    timestamp?: string;
    errors?: Array<string> | null;
};