// src/api/models/OrdenResponseApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { OrdenResponse } from './OrdenResponse';
export type OrdenResponseApiResponse = {
    success?: boolean;
    data?: OrdenResponse;
    message?: string | null;
    timestamp?: string;
    errors?: Array<string> | null;
};